Prevent repeated exits with socket activation
If an application shoves invalid input into the socket, the following
is probably invalid as well. systemd will, by default, send all
pending input to the restarted process, in case the service exited
with a non-zero code before. We don't really want that, as it might
cause wob to exit yet again. Therefore, systemd should just flush or
discard all pending data and move on.

Signed-off-by: Sefa Eyeoglu <contact@scrumplex.net>