Set StandardOutput in service file
If StandardInput is "socket" StandardOutput defaults to "inherit"
which is not what we want here ("inherit" causes stdout to be sent
back into the FIFO).

Setting it to "journal" gives the behaviour most will want/expect.

Ideally this would be set to the value of DefaultStandardOutput (see
man systemd-system.conf) but it doesn't look like there is an option
for that.